+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/resolveengine/graph/builder/DeselectVersionAction.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2017 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.api.internal.artifacts.ivyservice.resolveengine.graph.builder;
+
+import org.gradle.api.Action;
+import org.gradle.api.artifacts.ModuleIdentifier;
+
+class DeselectVersionAction implements Action<ModuleIdentifier> {
+    private final ResolveState resolveState;
+
+    DeselectVersionAction(ResolveState resolveState) {
+        this.resolveState = resolveState;
+    }
+
+    public void execute(ModuleIdentifier module) {
+        ComponentState previouslySelected = resolveState.getModule(module).clearSelection();
+        if (previouslySelected != null) {
+            for (NodeState configuration : previouslySelected.getNodes()) {
+                configuration.deselect();
+            }
+        }
+    }
+}
